moved get_table_string to utils in new file for text_formatting. Added doc string for it and imported it back into the Sketch module
